Apply suggestions from code review

+++ src/entity/enemy_stats.hpp
@@ -10,8 +10,8 @@ private:
     float health;
 
 public:
-    GETTER_SETTER(int, Health, health)
-    GETTER_SETTER(int, Attack_speed, attack_speed)
+    GETTER_SETTER(float, Health, health)
+    GETTER_SETTER(float, Attack_speed, attack_speed)
     GETTER_SETTER(float, Speed, speed)
     GETTER_SETTER(float, Size, size)
